refactor fragmented google auth endpoints
